# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PMS Hotel Management System API
  description: |
    Complete REST API documentation for the Property Management System.
    
    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header.
    
    ## Base URL
    - Development: http://localhost:3001/api
    - Production: https://api.yourhotel.com/api
    
    ## Error Codes
    - 400 - Bad Request
    - 401 - Unauthorized
    - 403 - Forbidden
    - 404 - Not Found
    - 500 - Internal Server Error
  version: 1.0.0
  contact:
    name: API Support
    email: support@grandview.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.grandview.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /settings/tax-rates:
    get:
      tags:
        - Settings
      summary: Get tax rates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tax rates
    post:
      tags:
        - Settings
      summary: Create tax rate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taxName
                - rate
              properties:
                taxName:
                  type: string
                taxCode:
                  type: string
                rate:
                  type: number
                taxType:
                  type: string
                country:
                  type: string
                isDefault:
                  type: boolean
      responses:
        '201':
          description: Tax rate created
  /settings/currencies:
    get:
      tags:
        - Settings
      summary: Get currencies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of currencies
    post:
      tags:
        - Settings
      summary: Create currency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currencyCode
                - currencyName
              properties:
                currencyCode:
                  type: string
                currencyName:
                  type: string
                symbol:
                  type: string
                exchangeRate:
                  type: number
                isDefault:
                  type: boolean
      responses:
        '201':
          description: Currency created
  /settings/languages:
    get:
      tags:
        - Settings
      summary: Get languages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of languages
    post:
      tags:
        - Settings
      summary: Create language
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - languageCode
                - languageName
              properties:
                languageCode:
                  type: string
                languageName:
                  type: string
                nativeName:
                  type: string
                isDefault:
                  type: boolean
      responses:
        '201':
          description: Language created
  /settings/payment-gateways:
    get:
      tags:
        - Settings
      summary: Get payment gateways
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of payment gateways
    post:
      tags:
        - Settings
      summary: Create payment gateway
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - gatewayName
              properties:
                gatewayName:
                  type: string
                gatewayType:
                  type: string
                configJson:
                  type: string
                isActive:
                  type: boolean
                isTestMode:
                  type: boolean
      responses:
        '201':
          description: Payment gateway created
  /settings/hotel:
    get:
      tags:
        - Settings
      summary: Get hotel information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Hotel information
    put:
      tags:
        - Settings
      summary: Update hotel information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotelName:
                  type: string
                address:
                  type: string
                city:
                  type: string
                country:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                website:
                  type: string
      responses:
        '200':
          description: Hotel information updated

  # ============================================================
  # NOTIFICATIONS
  # ============================================================
...